Feature/delete board

+++ service/column.go
@@ -2,25 +2,50 @@ package service
 
 import (
 	"context"
-	"github.com/google/uuid"
+	"errors"
+	"server/internal/board"
 	"server/internal/column"
+	userboardrole "server/internal/user_board_role"
 	"server/pkg/adapters/storage/entities"
+	"server/pkg/rbac"
+
+	"github.com/google/uuid"
+)
+
+var (
+	ErrPermissionDeniedToCreateColumn = errors.New("permission denied: can not create column")
+	ErrPermissionDeniedToDeleteColumn = errors.New("permission denied: can not delete the column")
 )
 
 type ColumnService struct {
-	colOps *column.Ops
+	colOps           *column.Ops
+	userBoardRoleOps *userboardrole.Ops
+	boardOps         *board.Ops
 }
 
-func NewColumnService(colOps *column.Ops) *ColumnService {
-	return &ColumnService{colOps: colOps}
+func NewColumnService(colOps *column.Ops, userBoardRoleOps *userboardrole.Ops, boardOps *board.Ops) *ColumnService {
+	return &ColumnService{colOps: colOps,
+		boardOps:         boardOps,
+		userBoardRoleOps: userBoardRoleOps}
 }
 
-func (s *ColumnService) CreateColumn(ctx context.Context, name string, boardID uuid.UUID, order uint) (*entities.Column, error) {
+func (s *ColumnService) CreateColumn(ctx context.Context, name string, boardID, userID uuid.UUID, order uint) (*entities.Column, error) {
 	col := &column.Column{
 		ID:       uuid.New(),
 		Name:     name,
 		BoardID:  boardID,
 		OrderNum: order,
+	}
+
+	//check to see board exists?
+	// check permission
+	role, err := s.userBoardRoleOps.GetUserBoardRole(ctx, userID, col.BoardID)
+	if err != nil {
+		return nil, ErrPermissionDeniedToCreateColumn
+	}
+
+	if !rbac.HasPermission(role, rbac.PermissionManageColumns) {
+		return nil, ErrPermissionDeniedToCreateColumn
 	}
 
 	if err := s.colOps.Create(ctx, col); err != nil {
@@ -64,7 +89,7 @@ func (s *ColumnService) GetMinOrderColumn(ctx context.Context, boardID uuid.UUID
 	}, nil
 }
 
-func (s *ColumnService) CreateColumns(ctx context.Context, columns []entities.Column) ([]entities.Column, error) {
+func (s *ColumnService) CreateColumns(ctx context.Context, columns []entities.Column, userID uuid.UUID) ([]entities.Column, error) {
 	colModels := make([]column.Column, len(columns))
 	for i, col := range columns {
 		colModels[i] = column.Column{
@@ -73,6 +98,24 @@ func (s *ColumnService) CreateColumns(ctx context.Context, columns []entities.Co
 			BoardID:  col.BoardID,
 			OrderNum: col.OrderNum,
 		}
+	}
+	//check to see board exists?
+	b, err := s.boardOps.GetBoardByID(ctx, colModels[0].BoardID)
+
+	if err != nil {
+		return nil, err
+	}
+	if b == nil {
+		return nil, board.ErrBoardNotFound
+	}
+	// check permission
+	role, err := s.userBoardRoleOps.GetUserBoardRole(ctx, userID, colModels[0].BoardID)
+	if err != nil {
+		return nil, ErrPermissionDeniedToCreateColumn
+	}
+
+	if !rbac.HasPermission(role, rbac.PermissionManageColumns) {
+		return nil, ErrPermissionDeniedToCreateColumn
 	}
 
 	createdCols, err := s.colOps.CreateColumns(ctx, colModels)
@@ -92,6 +135,25 @@ func (s *ColumnService) CreateColumns(ctx context.Context, columns []entities.Co
 	return createdEntities, nil
 }
 
-func (s *ColumnService) DeleteColumn(ctx context.Context, columnID uuid.UUID) error {
+func (s *ColumnService) DeleteColumn(ctx context.Context, columnID, userID uuid.UUID) error {
+
+	// get column if exists
+	col, err := s.colOps.GetColumnByID(ctx, columnID)
+	if err != nil {
+		return err
+	}
+	if col == nil {
+		return column.ErrColumnNotFound
+	}
+
+	// check permission
+	role, err := s.userBoardRoleOps.GetUserBoardRole(ctx, userID, col.BoardID)
+	if err != nil {
+		return ErrPermissionDeniedToDelete
+	}
+
+	if !rbac.HasPermission(role, rbac.PermissionManageColumns) {
+		return ErrPermissionDeniedToDelete
+	}
 	return s.colOps.Delete(ctx, columnID)
 }
